requirements-agent.txt
# Requirements for the LiveKit voice agent
livekit-agents>=1.3.0
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-silero
python-dotenv
duckduckgo-search
aiohttp
